Demonstrate feature #15 TS Decorators. Decorators are an experimental feature and may change in future TypeScript releases.
Added RequireRole decorator that we can attach to any
express route method to verify if the user's role is authorized to do the specific action.
Added few more decorators like error handler, logger, but these not used yet.
Enabled experimentalDecorators in tsconfig.
Update the dev script to clear the console when something changes in code.